Changes for NetBSD

+++ ex_cmdsub.c
@@ -1065,7 +1065,7 @@ addmac(char *src,char *dest,char *dname)
 		 * linefeed, and escape, he can screw himself. This is
 		 * so weird I don't bother to check for it.
 		 */
-		if ((isalpha(src[0]) && src[1]) || any(src[0],":"))
+		if ((isalpha((int)src[0]) && src[1]) || any(src[0],":"))
 			error("Too dangerous to map that");
 		/*
 		 * If the src were null it would cause the dest to
